cloudito-recognition-daemon/requirements.txt
numpy==1.18.2
opencv-contrib-python==3.4.2.17
opencv-python-headless==4.2.0.32
PyYAML==5.1.2
grpcio==1.28.1  
grpcio-tools==1.28.1 
mysql-connector==2.2.9
requests==2.23.0